feat(leaderboard): store in database

+++ crunch_titles/_repository.py
@@ -1,12 +1,43 @@
 from abc import ABC, abstractmethod
 from logging import Logger
 from textwrap import dedent
-from typing import Callable, Dict, List, Optional, Set, Tuple
+from typing import Callable, Dict, List, Optional, Set, Tuple, cast
 
 from tqdm.auto import tqdm
 
 from crunch_titles._database import Database, to_column_names, to_table_name
-from crunch_titles._model import Competition, CompetitionId, CompetitionName, Crunch, CrunchId, CrunchTarget, CrunchTargetId, Leaderboard, LeaderboardDefinition, LeaderboardDefinitionId, LeaderboardId, Payout, PayoutId, PayoutRecipient, Phase, PhaseId, Position, Round, RoundId, Target, TargetId, Title, TitlePosition, TitlePositionBody, User, UserId
+from crunch_titles._model import (
+    Competition,
+    CompetitionId,
+    CompetitionName,
+    Crunch,
+    CrunchId,
+    CrunchTarget,
+    CrunchTargetId,
+    Leaderboard,
+    LeaderboardDefinition,
+    LeaderboardDefinitionId,
+    LeaderboardId,
+    Payout,
+    PayoutId,
+    PayoutRecipient,
+    Phase,
+    PhaseId,
+    Position,
+    Round,
+    RoundId,
+    Target,
+    TargetId,
+    Title,
+    TitleLeaderboard,
+    TitleLeaderboardBody,
+    TitleLeaderboardId,
+    TitlePosition,
+    TitlePositionBody,
+    TitlePositionId,
+    User,
+    UserId
+)
 from crunch_titles._utility import group_by, to_dict
 
 
@@ -15,11 +46,15 @@ class Repository(ABC):
     @abstractmethod
     def find_all_competitions(self) -> List[Competition]:
         ...
+    
+    @abstractmethod
+    def find_competition_by_id(self, id: CompetitionId) -> Competition:
+        ...
 
     @abstractmethod
     def find_competition_by_name(self, name: CompetitionName) -> Competition:
         ...
-
+        
     @abstractmethod
     def find_user_by_id(self, id: UserId) -> User:
         ...
@@ -69,11 +104,23 @@ class Repository(ABC):
         ...
 
     @abstractmethod
-    def delete_title_positions_by_competition_and_year(self, competition: Competition, year: int):
+    def create_title_leaderboard(self, body: TitleLeaderboardBody) -> TitleLeaderboard:
         ...
 
     @abstractmethod
-    def create_title_position(self, body: TitlePositionBody):
+    def find_all_title_leaderboards(self) -> List[TitleLeaderboard]:
+        ...
+
+    @abstractmethod
+    def delete_title_leaderboard_by_competition_and_year(self, competition: Competition, year: int):
+        ...
+
+    @abstractmethod
+    def find_all_title_positions_by_leaderboard(self, leaderboard: TitleLeaderboard) -> List[TitlePosition]:
+        ...
+
+    @abstractmethod
+    def create_title_position(self, body: TitlePositionBody) -> TitlePosition:
         ...
 
     @abstractmethod
@@ -84,6 +131,7 @@ class Repository(ABC):
 class LoadEverythingRepository(Repository):
 
     _competitions: List[Competition]
+    _competition_by_id: Dict[CompetitionId, Competition]
     _competition_by_name: Dict[CompetitionName, Competition]
     _user_by_id: Dict[UserId, User]
     _user_by_login: Dict[str, User]
@@ -97,6 +145,8 @@ class LoadEverythingRepository(Repository):
     _position_by_leaderboard_id: Dict[LeaderboardId, List[Position]]
     _payouts_by_competition_id: Dict[CompetitionId, List[Payout]]
     _payout_recipients_by_payout_id: Dict[PayoutId, List[PayoutRecipient]]
+    _title_leaderboard_by_competition_id_and_year: Dict[Tuple[CompetitionId, int], TitleLeaderboard]
+    _title_positions_by_leaderboard_id: Dict[TitleLeaderboardId, List[TitlePosition]]
 
     def __init__(
         self,
@@ -119,6 +169,11 @@ class LoadEverythingRepository(Repository):
             self._competitions = self._database.competition.query_many_objects(
                 Competition,
                 where="`visibility` = 'PUBLIC' AND NOT `external`",
+            )
+
+            self._competition_by_id = to_dict(
+                self._competitions,
+                key=lambda row: row["id"],
             )
 
             self._competition_by_name = to_dict(
@@ -246,6 +301,18 @@ class LoadEverythingRepository(Repository):
                 key=lambda row: row["payout_id"],
             )
 
+        def _load_title_leaderboards():
+            self._title_leaderboard_by_competition_id_and_year = to_dict(
+                self._database.competition.query_many_objects(TitleLeaderboard),
+                key=lambda row: (row["competition_id"], row["year"]),
+            )
+
+        def _load_title_positions():
+            self._title_positions_by_leaderboard_id = group_by(
+                self._database.competition.query_many_objects(TitlePosition),
+                key=lambda row: row["leaderboard_id"],
+            )
+
         methods: List[Callable[[], None]] = [
             _load_competitions,
             _load_users,
@@ -258,6 +325,8 @@ class LoadEverythingRepository(Repository):
             _load_leaderboards,
             _load_positions,
             _load_paid_checkpoint_payouts,
+            _load_title_leaderboards,
+            _load_title_positions,
         ]
 
         for method in tqdm(methods, unit="method", miniters=1):
@@ -275,6 +344,9 @@ class LoadEverythingRepository(Repository):
 
     def find_all_competitions(self) -> List[Competition]:
         return self._competitions
+
+    def find_competition_by_id(self, id: CompetitionId) -> Competition:
+        return self._competition_by_id[id]
 
     def find_competition_by_name(self, name: CompetitionName) -> Competition:
         return self._competition_by_name[name]
@@ -315,26 +387,90 @@ class LoadEverythingRepository(Repository):
     def find_all_payout_recipients(self, payout: Payout) -> List[PayoutRecipient]:
         return self._payout_recipients_by_payout_id.get(payout["id"]) or []
 
-    def delete_title_positions_by_competition_and_year(self, competition: Competition, year: int):
+    def find_all_title_leaderboards(self) -> List[TitleLeaderboard]:
+        return list(self._title_leaderboard_by_competition_id_and_year.values())
+
+    def create_title_leaderboard(self, body: TitleLeaderboardBody) -> TitleLeaderboard:
+        id = self._database.competition.insert_object(
+            to_table_name(TitleLeaderboard),
+            body,
+        )
+
+        leaderboard: TitleLeaderboard = {
+            "id": cast(TitleLeaderboardId, id),
+            **body,
+        }
+
+        self._title_leaderboard_by_competition_id_and_year[(body["competition_id"], body["year"])] = leaderboard
+
+        return leaderboard
+
+    def delete_title_leaderboard_by_competition_and_year(self, competition: Competition, year: int):
+        leaderboard = self._database.competition.query_first_object(
+            TitleLeaderboard,
+            where=f"`competition_id` = %s AND `year` = %s",
+            params=(
+                competition["id"],
+                year,
+            ),
+        )
+
+        if not leaderboard:
+            return
+
+        cache_key = (competition["id"], year)
+        if cache_key in self._title_leaderboard_by_competition_id_and_year:
+            del self._title_leaderboard_by_competition_id_and_year[cache_key]
+
+        leaderboard_id = leaderboard["id"]
+        if leaderboard_id in self._title_positions_by_leaderboard_id:
+            del self._title_positions_by_leaderboard_id[leaderboard_id]
+
         self._database.competition.insert(
             f"""
                 DELETE FROM
                     `{to_table_name(TitlePosition)}`
                 WHERE
-                    `competition_id` = %s
-                    AND `year` = %s
+                    `leaderboard_id` = %s
             """,
-            (
-                competition["id"],
-                year,
+            params=(
+                leaderboard_id,
             )
         )
 
-    def create_title_position(self, body: TitlePositionBody):
-        self._database.competition.insert_object(
-            to_table_name(TitlePosition),
-            body
+        self._database.competition.insert(
+            f"""
+                DELETE FROM
+                    `{to_table_name(TitleLeaderboard)}`
+                WHERE
+                    `id` = %s
+            """,
+            params=(
+                leaderboard_id,
+            )
         )
+
+    def find_all_title_positions_by_leaderboard(self, leaderboard: TitleLeaderboard) -> List[TitlePosition]:
+        return self._title_positions_by_leaderboard_id.get(leaderboard["id"]) or []
+
+    def create_title_position(self, body: TitlePositionBody) -> TitlePosition:
+        id = self._database.competition.insert_object(
+            to_table_name(TitlePosition),
+            body,
+        )
+
+        position: TitlePosition = {
+            "id": cast(TitlePositionId, id),
+            **body,
+        }
+
+        leaderboard_id = position["leaderboard_id"]
+        if leaderboard_id not in self._title_positions_by_leaderboard_id:
+            self._title_positions_by_leaderboard_id[leaderboard_id] = []
+
+        self._title_positions_by_leaderboard_id[leaderboard_id].append(position)
+
+        return position
 
     def set_title_for_users(self, title: Title, user_ids: Set[UserId]) -> None:
         if not len(user_ids):
@@ -350,9 +486,11 @@ class LoadEverythingRepository(Repository):
                     `title` = %s
                 WHERE
                     `id` IN ({user_id_placeholders})
+                    AND `title` < %s
             """,
-            (
+            params=(
                 title,
                 *user_ids,
+                title,
             )
         )
